drf/v2/requirements.txt
djangorestframework == 3.15.2
requests == 2.32.3